Refactor panelManager unit tests by breaking it into individual files
